+++ src/main/java/com/huestew/studio/tools/PopulateTool.java
@@ -14,7 +14,7 @@ public class PopulateTool implements Tool {
 	@Override
 	public void doAction(Event event, LightTrack lightTrack, KeyFrame keyFrame, int timestamp, double normalizedY) {
 		if (event.getEventType() == MouseEvent.MOUSE_CLICKED) {
-			lightTrack.addKeyFrame(new KeyFrame(timestamp, new LightState(Color.green, (short) (255*normalizedY), (short) 0)));
+			lightTrack.addKeyFrame(new KeyFrame(timestamp, new LightState(Color.green, (int) (255 * normalizedY), 0)));
 		}
 	}
 }
